feat(breakpoints): DLT-2463 migrate doc site to new breakpoints
