+++ .npmrc
@@ -1,2 +1,3 @@
 publish-branch=production
 save-prefix=""
+@dialpad:registry=https://npm.pkg.github.com
