+++ apps/dialtone-documentation/package.json
@@ -51,6 +51,7 @@
     "@dialpad/dialtone-icons": "workspace:*",
     "@dialpad/dialtone-tokens": "workspace:*",
     "@dialpad/dialtone-vue": "workspace:^3",
+    "@dialpad/postcss-responsive-variations": "1.1.5",
     "@docsearch/js": "^3.5.1",
     "@originjs/vite-plugin-commonjs": "^1.0.3",
     "@toycode/markdown-it-class": "^1.2.4",
@@ -68,6 +69,7 @@
     "date-fns": "^4.1.0",
     "markdown-it": "^13.0.2",
     "markdown-it-anchor": "^8.6.7",
+    "postcss": "8.5.3",
     "prettier": "3.3.2",
     "prismjs": "^1.29.0",
     "vite-svg-loader": "^5.1.0",
